feat:_app.js refactor layout
