+++ pages/About.jsx
@@ -1,6 +1,8 @@
 import React from "react";
+import Container from "../components/Container";
+
 const About = () => {
-  return <div>About</div>;
+  return <p>About</p>;
 };
 
 export default About;
