+++ pages/_app.js
@@ -1,7 +1,23 @@
-import '../styles/globals.css'
+import { Layout } from "../components/Layout";
+import Head from "next/head";
+import "../styles/globals.css";
+import Container from "../components/Container";
 
 function MyApp({ Component, pageProps }) {
-  return <Component {...pageProps} />
+  return (
+    <>
+      <Head>
+        <title>Create Next App</title>
+        <meta name="description" content="Generated by create next app" />
+        <link rel="icon" href="/favicon.ico" />
+      </Head>
+      <Layout>
+        <Container>
+          <Component {...pageProps} />;
+        </Container>
+      </Layout>
+    </>
+  );
 }
 
-export default MyApp
+export default MyApp;
